added theme styles
i didnt want to commit the new changes until we met in person just in case it broke something with the databases. i had to change and delete a lot of code that seems redundant but i'm not sure if it broke anything yet. 

This file has separated css files for each webpage. i plan on looking over it again to see if i can reduce the number of pages by combining styles